Add config file to train on Cityscapes
Here is expected the Cityscapes dataset in the following manner under the data directory:
SOLO
├── mmdet
├── tools
├── configs
├── data
│   ├── cityscapes
│   │   ├── annotations
│   │   ├── train
│   │   ├── val

Where as noted in docs/INSTALL.md, all CS images for each split are in just one folder (be it in symlink or not). Here the annotations are named the same from the COCO transformation script.